Had to backout the -c option of esl-sfetch
esl-sfetch was failing when trying to get the sub sequences from pfamseq
using the -c option. This is a bug with the old version of esl-sfetch,
but we need to be able to get around it. So using a substring in perl
instead, so that we don't need to fix all old binary versions.

git-svn-id: https://xfamsvn.ebi.ac.uk/svn/code/trunk/PfamScripts@6399 98333f76-0502-11df-81af-936bbf3a1e26