This code has been the most heavily module impacted by the changes to use non-bioperl based objects to represent the full alignments. Most of the changes involve determining which type of object has been passed to the QC checks and then taking the appropriate action to deal with the array or Bio::Seq object.
I have also make some substantial changes to the way that
the signal peptide QC was working. Instead of making a 
hash of every sequence in the alignment (admittedly it was
taking the N-terminal domain), I now only look for hits
that fall with the first 120 amino acids. This makes a 
signficiant gain for many families, with only 296 signal 
peptides out of 2.2 million ending after that start position.


git-svn-id: https://xfamsvn.ebi.ac.uk/svn/code/trunk/PfamLib@6783 98333f76-0502-11df-81af-936bbf3a1e26