Disabled C style comments in the config files.
These are not useful to anyone and have cuased problems when generating
the robots.txt files. In order to stop google from crawling a url like
/family/PF0001/alignment the robots.txt file needs to have an entry like:

Disallow: /family/*/alignment

The problem with C style comments is that Config::General was commenting
out everything on the line up to an including /*/. This meant there was
no way to stop google from crawling those links.

git-svn-id: https://xfamsvn.ebi.ac.uk/svn/code/trunk/PfamBase@6851 98333f76-0502-11df-81af-936bbf3a1e26