New view process code!
This has changed from being method based to OO.  There are also a large number of changes
to reflect the changing nature of what we going to provide via the website. This
includes making RP alignments, not making trees or searching against the other datasets
(metagneomic and ncbi). HTML alignments will only be produced for aligments less than
5000 sequences. 



git-svn-id: https://xfamsvn.ebi.ac.uk/svn/code/trunk/PfamLib@7102 98333f76-0502-11df-81af-936bbf3a1e26