This is a test of the email notification hook in the Pfam SVN repository.
Currently this is sending mail to JGT, JD, SWB, RF and EN. More names can
be added to this list by editing the post-commit hook in the repository.

It's also restricted to just the Rfam module. If we want notifications
from the whole repository, that's only a tweak to the post-commit script
too.


git-svn-id: https://xfamsvn.ebi.ac.uk/svn/code/trunk/Rfam@7511 98333f76-0502-11df-81af-936bbf3a1e26