Added some cleaning up to SecondaryStructure.pm
Significant changes to Tree.pm. FastTree will not work whenever the sequence identifiers are the same, so we can't build a tree on the seed where we have replaced the NSE with the species, as there are many families where they have more than one sequence from a single species. Therefore, we need to search and replace on the tree to replace all the sequence accessions with species; there is now a subroutine (specifyTree) which does this.

All trees loaded into the database


git-svn-id: https://xfamsvn.ebi.ac.uk/svn/code/trunk/Rfam@8054 98333f76-0502-11df-81af-936bbf3a1e26