Added subtypes for Pfam-A accession and ID.
There was a problem with clci, which uses both Bio::Pfam::Family::DESC and
Bio::Pfam::Family::PfamA. Trying to use B::P::F::PfamA caused a moose exception
with an error message about trying to define a subtype ("PfamAcc") that was
already defined.

The PfamAcc subtype wasn't explicitly defined in B::P::F::DESC, but was implied
by the "AC" accessor, which seems to have implicitly created a subtype. Trying
then to use B::P::F::PfamA then triggers the exception. 

Adding the subtypes (with different names, "DescPfamAcc") here is a bit of a
hack. We should probably collect up all of the subtypes and put them in a
separate package that can be included by everything that needs them.


git-svn-id: https://xfamsvn.ebi.ac.uk/svn/code/trunk/PfamLib@8075 98333f76-0502-11df-81af-936bbf3a1e26