Clean up 3D table
Adds a little text to the header, which I may add more to later if
needed. Also improves the table display a little. It is cleaner to use
', ' as the separator as that way the table can resize and not be very
wide. Additionally, links to the PDB structures were added and the names
of links were changed.